move bintokens to aici folder
